Added reference to the sketch for PCB_mcu_rev1. I realized after a short walk that it's better to continue tomorrow with the contents of the freecad readme. I will tomorrow also need to review the sketch referenced in this new addition to readme commited in this commit.
Then ask for reviews to maybe clarify possible confusion or insanity of
the repo.

But I feel I then need to call quits on this project; only thing
reminding is the fun/rewarding things of assembling, adding assembly instructions
and gifting it/delivering to the end-user.

But now food and sleep.